Refactor main()
For cleanup.
