+++ lib/main.dart
@@ -1,48 +1,20 @@
 //@dart=2.9
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
+import 'package:website/components/FirebaseInit.dart';
+import 'package:website/components/licences.dart';
 
 //Internal Packages
 import 'dom.dart';
 import 'components/theme_changer.dart';
-import 'components/models/ip_location.dart';
 
 void main() async {
   //Licenses
-  //OFL for Google Fonts
-  LicenseRegistry.addLicense(() async* {
-    final license = await rootBundle.loadString('assets/licenses/OFL.txt');
-    yield LicenseEntryWithLineBreaks(['google_fonts'], license);
-  });
-  //Unsplash license for picture
-  LicenseRegistry.addLicense(() async* {
-    final unsplashLicense =
-        await rootBundle.loadString("assets/licenses/Unsplash.txt");
-    yield LicenseEntryWithLineBreaks(["unsplash"], unsplashLicense);
-  });
+  await Licences.load();
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp();
-  UserCredential userCredential =
-      await FirebaseAuth.instance.signInAnonymously();
-  FirebaseAnalytics analytics = FirebaseAnalytics();
-  analytics.setUserId(userCredential.user.uid);
-  IPLocation location = await IPLocation.getLocation();
-  analytics.setUserProperty(name: "IP", value: location.query);
-  analytics.setUserProperty(name: "State", value: location.regionName);
-  analytics.setUserProperty(name: "City", value: location.city);
-  analytics.setUserProperty(name: "ISP", value: location.isp);
-  analytics.setUserProperty(name: "Zip", value: location.zip);
-  var now = DateTime.now();
-  analytics.setUserProperty(
-      name: "lastTimeAccessed",
-      value: DateTime(now.year, now.month, now.day, now.hour, now.minute)
-          .toString());
-  analytics.logAppOpen();
+  //Initialize Firebase
+  await FirebaseInit.init();
   runApp(MyApp());
 }
